cli_tool/requirements.txt
click>=8.0.0
requests>=2.28.0
pyyaml>=6.0
rich>=13.0.0
httpx>=0.24.0
pydantic>=2.0.0
fastapi==0.104.1
uvicorn==0.24.0